+++ software/remote/sources/hardware/bq27441.h
@@ -83,7 +83,7 @@ public:
     Q_INVOKABLE void begin();
     Q_INVOKABLE int getVoltage();
     Q_INVOKABLE int getFullChargeCapacity();
-    Q_INVOKABLE int16_t  getAverageCurrent();
+    Q_INVOKABLE int getAverageCurrent();
     Q_INVOKABLE int getAveragePower();
     Q_INVOKABLE int getStateOfCharge();
     Q_INVOKABLE int16_t  getInternalTemperatureC();           // Result in 0.1 Celsius
